+++ src/menu.rs
@@ -221,6 +221,7 @@ const VIEW_EDITOR: &[Item] = &[
     Item::leaf("menu.item.view.inlay_hints", "view.inlay_hints", ""),
     Item::leaf("menu.item.view.spellcheck", "view.spellcheck", ""),
     SEP,
+    Item::leaf("menu.item.view.auto_pair", "view.auto_pair", ""),
     Item::leaf("menu.item.view.trim_on_save", "view.trim_on_save", ""),
     Item::leaf("menu.item.view.final_newline_on_save", "view.final_newline_on_save", ""),
     SEP,